// src/lib/types/colors.ts
/* SPDX-License-Identifier: MIT
Vegan Activists (VeganActivists@proton.me) */

/**
 * Color Codes Defined By Discord
 */
enum Colors {
  DEFAULT = 0x00_00_00,
  WHITE = 0xff_ff_ff,
  AQUA = 0x1a_bc_9c,
  GREEN = 0x2e_cc_71,
  BLUE = 0x34_98_db,
  YELLOW = 0xff_ff_00,
  PURPLE = 0x9b_59_b6,
  LUMINOUS_VIVID_PINK = 0xe9_1e_63,
  GOLD = 0xf1_c4_0f,
  ORANGE = 0xe6_7e_22,
  RED = 0xe7_4c_3c,
  GREY = 0x95_a5_a6,
  NAVY = 0x34_49_5e,
  DARK_AQUA = 0x11_80_6a,
  DARK_GREEN = 0x1f_8b_4c,
  DARK_BLUE = 0x20_66_94,
  DARK_PURPLE = 0x71_36_8a,
  DARK_VIVID_PINK = 0xad_14_57,
  DARK_GOLD = 0xc2_7c_0e,
  DARK_ORANGE = 0xa8_43_00,
  DARK_RED = 0x99_2d_22,
  DARK_GREY = 0x97_9c_9f,
  DARKER_GREY = 0x7f_8c_8d,
  LIGHT_GREY = 0xbc_c0_c0,
  DARK_NAVY = 0x2c_3e_50,
  BLURPLE = 0x72_89_da,
  GREYPLE = 0x99_aa_b5,
  DARK_BUT_NOT_BLACK = 0x2c_2f_33,
  NOT_QUITE_BLACK = 0x23_27_2a,
}

export default Colors;